Initial Laravel Lumen Commit
This includes a docker-compose file, and three development tables for
troubleshooting.

So far, hopefully the only thing you have to do to set it up is
`docker-compose up -d`. You'll be able to go inside the container with
`docker-compose exec app bash`. Inside the container, you can do things
like `php artisan migrate` to migrate the data to the sql server

So far, there's no functionality, but there's a lot of potential