Add basic database structure
All viewers of the site are going to be associated with a token. The
token may or may not be associated with a user. The token is associated
with a blog post (to signify a unique page count). Blog posts are
associated with users (to signify them being an editor). Comments are
associated with blog posts. And login attempts are associated with
a token, user, and an ip address