Add Reactor helper
Reactor is a "as minimal as possible" implementation of reactivity
This is eventually going to completely replace Vue for internal use. Say
goodbye to 80kb of bloatware :)

It doesn't do everything Vue does. It doesn't do computed or methods.
Nor does it do single file components. It's not designed to, and no
features will be implemented to replicate Vue in this regard. What it
does try to accomplish is a minimal way to detect changes to variables